Update noderedtest.py
I think we can try the third option for NodeRed, as we will know the port that NodeRed reports from in minimal modBus. 
We will have to install MinimalModbus as well. 
Some are saying that NodeRed is on port 1880 (but that will change on each DL design)

There may be some changes to the Wh to convert it to MWh i.e. divide by 1,000,000